requirements.txt
asgiref==3.6.0
Django==4.2
django-cors-headers==3.14.0
django-crontab==0.7.1
django-filter==23.2
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
